# Speed up RSpec test (closes #1932)