--- a/spec/api/v2/public/markets_spec.rb
+++ b/spec/api/v2/public/markets_spec.rb
@@ -485,19 +485,24 @@ describe API::V2::Public::Markets, type: :request do
 
     it 'gets trades by page and limit', clean_database_with_truncation: true do
       # pending 'because of database_cleaner'
+      # binding.pry
+
       create(:trade, :btcusd, bid: bid, created_at: 6.hours.ago)
 
+      # binding.pry
       get "/api/v2/public/markets/#{market}/trades", limit: 1, page: 1, order_by: 'asc'
 
       expect(response).to be_success
       expect(response.headers.fetch('Total')).to eq '3'
-      # expect(JSON.parse(response.body).first['id']).to eq 1
+
+      # binding.pry
+      expect(JSON.parse(response.body).first['id']).to eq 1
 
       get "/api/v2/public/markets/#{market}/trades", market: 'btcusd', limit: 1, page: 2, order_by: 'asc'
 
       expect(response).to be_success
       expect(response.headers.fetch('Total')).to eq '3'
-      # expect(JSON.parse(response.body).first['id']).to eq 2
+      expect(JSON.parse(response.body).first['id']).to eq 2
     end
 
     it 'validates market param' do