--- a/spec/models/payment_address_spec.rb
+++ b/spec/models/payment_address_spec.rb
@@ -6,8 +6,11 @@ describe PaymentAddress do
     let(:member)  { create(:member, :level_3) }
     let!(:account) { member.get_account(:btc) }
 
-    it 'generate address after commit' do
-      pending 'because of database_cleaner'
+    after do
+      DatabaseCleaner.strategy = :truncation
+    end
+
+    it 'generate address after commit', clean_database_with_truncation: true do
       AMQPQueue.expects(:enqueue)
                .with(:deposit_coin_address, { account_id: account.id }, { persistent: true })
       account.payment_address