--- a/spec/serializers/serializers/event_api/order_canceled_spec.rb
+++ b/spec/serializers/serializers/event_api/order_canceled_spec.rb
@@ -34,6 +34,7 @@ describe Serializers::EventAPI::OrderCanceled do
   before { OrderAsk.any_instance.expects(:updated_at).returns(canceled_at).at_least_once }
 
   before do
+    DatabaseCleaner.clean
     EventAPI.expects(:notify).with('market.btcusd.order_created', anything).once
     EventAPI.expects(:notify).with('market.btcusd.order_canceled', {
       id:                       1,
@@ -59,13 +60,15 @@ describe Serializers::EventAPI::OrderCanceled do
     }).once
   end
 
-  it 'publishes event' do
-    pending 'because of database_cleaner'
+  after do
+    DatabaseCleaner.strategy = :truncation
+  end
+
+  it 'publishes event', clean_database_with_truncation: true do
     subject
     subject.transaction do
-      subject.state = Order::CANCEL
+      subject.update!(state: Order::CANCEL)
       subject.hold_account.unlock_funds(subject.locked)
-      subject.save!
     end
   end
 end