--- a/spec/serializers/serializers/event_api/order_created_spec.rb
+++ b/spec/serializers/serializers/event_api/order_created_spec.rb
@@ -32,6 +32,7 @@ describe Serializers::EventAPI::OrderCreated do
   before { OrderBid.any_instance.expects(:created_at).returns(created_at).at_least_once }
 
   before do
+    DatabaseCleaner.clean
     EventAPI.expects(:notify).with('market.btcusd.order_created', {
       id:                     1,
       market:                 'btcusd',
@@ -55,8 +56,11 @@ describe Serializers::EventAPI::OrderCreated do
     }).once
   end
 
-  it('publishes event') do
-    pending 'because of database_cleaner'
+  after do
+    DatabaseCleaner.strategy = :truncation
+  end
+
+  it('publishes event', clean_database_with_truncation: true) do
     subject
   end
 end