--- a/spec/serializers/serializers/event_api/order_updated_spec.rb
+++ b/spec/serializers/serializers/event_api/order_updated_spec.rb
@@ -34,6 +34,7 @@ describe Serializers::EventAPI::OrderUpdated, OrderAsk do
   before { OrderAsk.any_instance.expects(:updated_at).returns(updated_at).at_least_once }
 
   before do
+    DatabaseCleaner.clean
     EventAPI.expects(:notify).with('market.btcusd.order_created', anything)
     EventAPI.expects(:notify).with('market.btcusd.order_updated', {
       id:                      1,
@@ -61,8 +62,11 @@ describe Serializers::EventAPI::OrderUpdated, OrderAsk do
     }).once
   end
 
-  it 'publishes event' do
-    pending 'because of database_cleaner'
+  after do
+    DatabaseCleaner.strategy = :truncation
+  end
+
+  it 'publishes event', clean_database_with_truncation: true do
     subject.transaction do
       subject.update! \
         volume:         80,
@@ -106,6 +110,7 @@ describe Serializers::EventAPI::OrderUpdated, OrderBid do
   before { OrderBid.any_instance.expects(:updated_at).returns(updated_at).at_least_once }
 
   before do
+    DatabaseCleaner.clean
     EventAPI.expects(:notify).with('market.btcusd.order_created', anything)
     EventAPI.expects(:notify).with('market.btcusd.order_updated', {
       id:                      1,
@@ -133,8 +138,11 @@ describe Serializers::EventAPI::OrderUpdated, OrderBid do
     }).once
   end
 
-  it 'publishes event' do
-    pending 'because of database_cleaner'
+  after do
+    DatabaseCleaner.strategy = :truncation
+  end
+
+  it 'publishes event', clean_database_with_truncation: true do
     subject.transaction do
       subject.update! \
         volume:         12,