--- a/spec/serializers/serializers/event_api/trade_completed_spec.rb
+++ b/spec/serializers/serializers/event_api/trade_completed_spec.rb
@@ -65,6 +65,7 @@ describe Serializers::EventAPI::TradeCompleted, 'Event API' do
   before { Trade.any_instance.expects(:created_at).returns(completed_at).at_least_once }
 
   before do
+    DatabaseCleaner.clean
     EventAPI.expects(:notify).with('market.btcusd.order_created', anything).twice
     EventAPI.expects(:notify).with('market.btcusd.order_updated', anything).once
     EventAPI.expects(:notify).with('market.btcusd.order_completed', anything).once
@@ -90,8 +91,11 @@ describe Serializers::EventAPI::TradeCompleted, 'Event API' do
     }).once
   end
 
-  it 'publishes event' do
-    pending 'because of database_cleaner'
+  after do
+    DatabaseCleaner.strategy = :truncation
+  end
+
+  it 'publishes event', clean_database_with_truncation: true do
     subject
     expect(order_bid.reload.state).to eq 'done'
     expect(order_ask.reload.state).to eq 'wait'