Article: https://thoughtbot.com/blog/debugging-why-your-specs-have-slowed-down

## Top 5 slowest example groups:
* `API::V2::Management::Tools
    1.11 seconds average (1.11 seconds / 1 example) ./spec/api/v2/management/tools_spec.rb:4`
* `API::V2::Management::Operations
    1.02 seconds average (53.16 seconds / 52 examples) ./spec/api/v2/management/operations_spec.rb:4`
* `API::V2::Management::Withdraws
    0.92408 seconds average (21.25 seconds / 23 examples) ./spec/api/v2/management/withdraws_spec.rb:4`
* `API::V2::Management::Transfers
    0.90424 seconds average (25.32 seconds / 28 examples) ./spec/api/v2/management/transfers_spec.rb:4`
* `API::V2::Management::Deposits
    0.89089 seconds average (19.6 seconds / 22 examples) ./spec/api/v2/management/deposits_spec.rb:4`

## Timing

Finished in 6 minutes 27 seconds (files took 2.79 seconds to load)